crypto: fix VC++ warning C4244

+++ src/node_crypto.cc
@@ -761,7 +761,7 @@ void SecureContext::SetOptions(const FunctionCallbackInfo<Value>& args) {
     return sc->env()->ThrowTypeError("Bad parameter");
   }
 
-  SSL_CTX_set_options(sc->ctx_, args[0]->IntegerValue());
+  SSL_CTX_set_options(sc->ctx_, static_cast<long>(args[0]->IntegerValue()));
 }
